switch to using a u32 for indices
shouldn't matter I think. Strings and Vec<T>s are 24 stack bytes. On a
64-bit arch I think having a struct with a String, Vec, and 2 `u16`s
plus padding will be the same size as one with 2 `u32`s instead

the page IMDB has 286,361 parents (!)

+++ links/src/link_state/entry.rs
@@ -9,8 +9,8 @@
 pub struct Entry {
     pub title: String,
     neighbors: Vec<u32>,
-    last_parent: u16,
-    first_child: u16,
+    last_parent: u32,
+    first_child: u32,
 }
 
 use link_state::link_data::IndexedEntry;
